added tests

+++ src/main/java/com/ss/ita/kata/implementation/anastasia8755/SixImpl.java
@@ -1,11 +1,11 @@
 package com.ss.ita.kata.implementation.anastasia8755;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.ArrayList;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 public class SixImpl implements com.ss.ita.kata.Six {
     @Override
@@ -14,7 +14,7 @@ public class SixImpl implements com.ss.ita.kata.Six {
             //n^3 + (n-1)^3 + ... + 1^3 = m if m=0 => cubeCount=0
             if (m > 0) {
                 long currCubeVol = (long) Math.pow(cubeCount + 1, 3);
-                m = (long) m - currCubeVol;
+                m = m - currCubeVol;
 
             } else if (m == 0) {
                 return cubeCount;
